Fix country & region filtering
There is currently an issue with keyword filtering with either multiple
filters or none. Multiple will create a query looking for all of the items
rather than any. Meaning empty queries that should have results. While
none will return an error.

By way of example. A search parameter with the following keyword filters:
```
{'family_geography': ['AUS', 'BRN' ...]}
```

Produces a yql query like so:

```
and (family_geography contains "AUS")
and (family_geography contains "BRN")
and ...
```

This functionality is important for the backend region searches.

A related issue is the scenario where filters are:
```
{'family_geography': []}
```

This currently results in a query error, because the generated yql would
be:

```
 and ()
```