Add modular `YQLBuilder` with parity to `build_yql`
This aims to move from a long function that to a modular class. The main
aim here is to have a more modular approach to building the query. This
is in order to make it easier to maintain, as we can make more isolated
changes and have more granular tests. Using string templates also
contribute to the modularity as they likely can be more easily repurposed
in the future for different queries then string concatenation.

I've included in this commit the scrappy little script used to confirm
the query matches the former one exactly. This will then be removed in a
follow up commit