Make filters exclusive to each other
Previously, keyword filters where joined by an OR. This made sense for
searching within each keyword filter, which might have multiple geographies
or categories to include, but was not what we wanted when combining with
other filters. The result being that searching for a country and a category
would return a full list of everything from both, rather than just the
intersection between the two.